requirements.txt
Django==2.2
psycopg2>=2.7,<3.0
Pillow==2.2.1
django-filter==2.1.0
django-rest-swagger==2.2.0
djangorestframework==3.9.2
Markdown==3.1
